device-management/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
pika==1.3.2
paho-mqtt==1.6.1
requests==2.31.0